+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.2.0] - 12.04.2021
+
+- Added ability to add spacing around the grid with `horizontalGridMargin` and `verticallGridMargin`
+- Added ability to customize horizontal and vertical spacing between grid items with `horizontalGridSpacing` and `verticalGridSpacing`
+
+### Migrating from version [1.1.0]
+- Replace `spacing` with `horizontalGridSpacing` and `verticalGridSpacing`
+
 ## [1.1.0] - 08.03.2021
 
 - Updated dependencies and package to null safety